requirements.txt
numpy
pandas
tqdm
Pillow
scikit-learn
matplotlib
seaborn
lightgbm